时隔许久的同步主仓库

+++ pm-project-backdoor/src/main/java/cn/edu/xjtu/stu/orangesoft/backdoor/pojo/User.java
@@ -12,7 +12,7 @@ public class User {
     private String UserName;
     private String UserPassword;
     private int RoleID;
-    private int Tel;
+    private long Tel;
     private String EMail;
     private String UserIconLocation;
     private String Sex;
